Renaming to match typescript conventions
