+++ odh-jupyter-trash-cleanup/src/ITrashEmptyResponse.ts
